Move AccessLogDao to package 'dao.access.log'
This move allows the DAO class and related query mapping object
to be colated in the same package, improves cohesion.
