+++ lobby-db-dao/src/main/java/org/triplea/lobby/server/db/dao/access/log/AccessLogRecord.java
@@ -1,4 +1,4 @@
-package org.triplea.lobby.server.db.data;
+package org.triplea.lobby.server.db.dao.access.log;
 
 import java.time.Instant;
 import lombok.AllArgsConstructor;
