+++ lobby-db-dao/src/test/java/org/triplea/lobby/server/db/dao/access/log/AccessLogRecordTest.java
@@ -1,4 +1,4 @@
-package org.triplea.lobby.server.db.data;
+package org.triplea.lobby.server.db.dao.access.log;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
